requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
black==21.9b0
boto3==1.18.61
botocore==1.21.61
celery==5.1.2
certifi==2021.10.8
cffi==1.14.6
chardet==3.0.4
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
coverage==6.0.2
cryptography==2.9.2
Django==3.2.8
django-celery-beat==2.2.1
django-cors-headers==3.10.0
django-debug-toolbar==3.2.2
django-filter==21.1
django-storages==1.12.1
django-timezone-field==4.2.1
djangorestframework==3.12.4
drf-jwt==1.13.4
drf-yasg==1.20.0
gunicorn==20.1.0
idna==2.8
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.2
jmespath==0.10.0
joblib==1.1.0
kombu==5.1.0
Markdown==3.3.4
MarkupSafe==2.0.1
mypy-extensions==0.4.3
numpy==1.21.3
packaging==21.0
pandas==1.3.4
pathspec==0.9.0
platformdirs==2.4.0
prompt-toolkit==3.0.20
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
regex==2021.10.8
requests==2.22.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
scikit-learn==1.0.1
scipy==1.7.1
sentry-sdk==1.4.3
six==1.16.0
sqlparse==0.4.2
threadpoolctl==3.0.0
tomli==1.2.1
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.25.11
vine==5.0.0
wcwidth==0.2.5
